+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-30T07:57:15.372Z",
+    "2026-03-30T08:19:08.752Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {